% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_force_layout.R
\name{build_force_layout}
\alias{build_force_layout}
\title{Build a force-directed layout using igraph if available.}
\usage{
build_force_layout(edges, nodes, layout_method = LAYOUT_METHOD)
}
\arguments{
\item{edges}{Data.frame returned by read_causalpath_sif.}

\item{nodes}{Character vector of node names in desired order.}

\item{layout_method}{igraph layout function name.}
}
\value{
A data.frame with node, x, y.
}
\description{
Build a force-directed layout using igraph if available.
}
\keyword{internal}
