% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_node_layout.R
\name{build_node_layout}
\alias{build_node_layout}
\title{Build node metadata and layout.}
\usage{
build_node_layout(edges, format_data = NULL,
  layout_method = LAYOUT_METHOD, detailed_view = FALSE)
}
\arguments{
\item{edges}{Data.frame returned by read_causalpath_sif.}

\item{format_data}{Format data list for modification labels.}

\item{layout_method}{igraph layout function name.}

\item{detailed_view}{Whether to show detailed modification labels.}
}
\value{
A data.frame with node, x, y, width, height.
}
\description{
Build node metadata and layout.
}
\keyword{internal}
