% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_format_path.R
\name{derive_format_path}
\alias{derive_format_path}
\title{Derive a format file path from the input path.}
\usage{
derive_format_path(input_path)
}
\arguments{
\item{input_path}{Path to the input file.}
}
\value{
Path for the format TSV file.
}
\description{
Derive a format file path from the input path.
}
\keyword{internal}
