% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_output_path.R
\name{derive_output_path}
\alias{derive_output_path}
\title{Derive output file names from the input path.}
\usage{
derive_output_path(input_path, new_ext)
}
\arguments{
\item{input_path}{Path to the input file.}

\item{new_ext}{Extension to use (including dot).}
}
\value{
File path with the extension replaced.
}
\description{
Derive output file names from the input path.
}
\keyword{internal}
