% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_node_box.R
\name{draw_node_box}
\alias{draw_node_box}
\title{Draw a labeled node box.}
\usage{
draw_node_box(x, y, width, height, label, fill_color = NODE_FILL)
}
\arguments{
\item{x}{Center x.}

\item{y}{Center y.}

\item{width}{Box width.}

\item{height}{Box height.}

\item{label}{Node label.}

\item{fill_color}{Fill color for the node box.}
}
\value{
NULL. Draws on the active device.
}
\description{
Draw a labeled node box.
}
\keyword{internal}
