% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mod_format.R
\name{get_mod_format}
\alias{get_mod_format}
\title{Get modification formatting for a node.}
\usage{
get_mod_format(node_id, mod_id, format_data)
}
\arguments{
\item{node_id}{Node identifier.}

\item{mod_id}{Modification identifier.}

\item{format_data}{Format data list.}
}
\value{
List with color and label or NULL if not found.
}
\description{
Get modification formatting for a node.
}
\keyword{internal}
