% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node_color.R
\name{get_node_color}
\alias{get_node_color}
\title{Get the node color with fallback to default.}
\usage{
get_node_color(node_id, format_data)
}
\arguments{
\item{node_id}{Node identifier.}

\item{format_data}{Format data list from parse_format_file.}
}
\value{
R color string.
}
\description{
Get the node color with fallback to default.
}
\keyword{internal}
