% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_capsule_utils.R
\name{measure_mod_capsule}
\alias{measure_mod_capsule}
\title{Measure capsule dimensions for a modification label.}
\usage{
measure_mod_capsule(label)
}
\arguments{
\item{label}{Text label for the capsule.}
}
\value{
List with width and height in user units.
}
\description{
Measure capsule dimensions for a modification label.
}
\keyword{internal}
