% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_format_file.R
\name{parse_format_file}
\alias{parse_format_file}
\title{Parse a format file to get node and modification colors.}
\usage{
parse_format_file(format_path)
}
\arguments{
\item{format_path}{Path to the format TSV file.}
}
\value{
A list with default_node_color, node_colors, mod_formats.
}
\description{
Parse a format file to get node and modification colors.
}
\keyword{internal}
