% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rgb_color.R
\name{parse_rgb_color}
\alias{parse_rgb_color}
\title{Parse an RGB string into an R color.}
\usage{
parse_rgb_color(rgb_text, default_color = DEFAULT_NODE_COLOR)
}
\arguments{
\item{rgb_text}{String with space-delimited RGB values.}

\item{default_color}{Color to return if parsing fails.}
}
\value{
R color string.
}
\description{
Parse an RGB string into an R color.
}
\keyword{internal}
