% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_utils.R
\name{pixels_to_user_x}
\alias{pixels_to_user_x}
\title{Convert pixel length to user units along the x-axis.}
\usage{
pixels_to_user_x(pixels)
}
\arguments{
\item{pixels}{Pixel length to convert.}
}
\value{
Numeric length in user coordinates.
}
\description{
Convert pixel length to user units along the x-axis.
}
\keyword{internal}
