% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_causalpath.R
\name{render_causalpath}
\alias{render_causalpath}
\title{Render the causal network on the active device.}
\usage{
render_causalpath(edges, format_data, layout_method = LAYOUT_METHOD,
  detailed_view = FALSE)
}
\arguments{
\item{edges}{Data.frame of edges.}

\item{format_data}{Format data list for coloring.}

\item{layout_method}{igraph layout function name.}

\item{detailed_view}{Whether to show detailed modification labels.}
}
\value{
NULL. Draws to the active device.
}
\description{
Render the causal network on the active device.
}
\keyword{internal}
