% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sif.R
\name{search_list_of_vectors}
\alias{search_list_of_vectors}
\title{Find row indices where target vectors contain query values.}
\usage{
search_list_of_vectors(query_values, target_values)
}
\arguments{
\item{query_values}{Character vector of query values.}

\item{target_values}{Character vector of target values (possibly delimited).}
}
\value{
List of index vectors per query.
}
\description{
Find row indices where target vectors contain query values.
}
\keyword{internal}
