% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_to_rect.R
\name{trim_to_rect}
\alias{trim_to_rect}
\title{Compute intersection of a line from center with a rectangle boundary.}
\usage{
trim_to_rect(x0, y0, width, height, x1, y1)
}
\arguments{
\item{x0}{Rectangle center x.}

\item{y0}{Rectangle center y.}

\item{width}{Rectangle width.}

\item{height}{Rectangle height.}

\item{x1}{Target point x.}

\item{y1}{Target point y.}
}
\value{
Numeric vector c(x, y) for the intersection point.
}
\description{
Compute intersection of a line from center with a rectangle boundary.
}
\keyword{internal}
