% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_utils.R
\name{value_to_rgb}
\alias{value_to_rgb}
\title{Convert a value into an RGB string using a blue-white-red scale.}
\usage{
value_to_rgb(value, value_min = -2, value_max = 2)
}
\arguments{
\item{value}{Numeric value to map.}

\item{value_min}{Minimum of the scale range.}

\item{value_max}{Maximum of the scale range.}
}
\value{
Space-delimited RGB string.
}
\description{
Convert a value into an RGB string using a blue-white-red scale.
}
\keyword{internal}
